Adds renderers tests cases to tests_parsers and renames
